Fixes for reset of vhost-user device
1) Call the backend::deactivate_device() unconditionally
from transport::deactivate_device().
Before the patch, it was called only if device activated.
But device can be in half-activated state and we need to
reset it again

2) Implement function reset_vhost_user(). For virtio-fs,
this is a message that executes reset at virtiofsd side.

3) Pass parameter to client::reset_vhost_user() about
how device should be reset: using reset_vhost_user, or
by sending message to vhost-user daemon to deconfigure
queues

4) Call client::reset_vhost_user() from virtio-fs
to reset the device

5) the client::reset_vhost_user() must not return
error immediately if communication with client failed.
We must clear remembered events always

Signed-off-by: Dmitry Skorodumov <skorodumov.dmitry@huawei.com>