devices: delete parameters in funtion `realize` for all devices
Delete parameters(except `bus`) in funtion `realize` for all devices,
and move them to function `new`.
This is the first step to unify the use of the `realize` function which
will be moved to `device` trait.

Note: the `bus` parameter in function `realize` was not deleted
as it will be removed in subsequent patches.

Signed-off-by: liuxiangdong <liuxiangdong5@huawei.com>