test: add `address_space_init` function in devices/virtio crate
Many tests require the creation of the address space.
Use the unified `address_space_init` function to complete this task.

In rust, a type under `#[cfg(test)]` exported in crate A can not be used
for another test under `#[cfg(test)]` in crate B. So we can not add this
function by `#[cfg(test)]` in address_space crate.

Because currently only virtio and devices crates will use it, create function
`address_space_init` for each crate.

Signed-off-by: liuxiangdong <liuxiangdong5@huawei.com>