address_space: fix searching ioeventfd error
The address that ioeventfd listens to is not unique. For example,
virtio-mmio device will listen to `QueueNotify` register(0x050).
This register will receive a value to notify this virtio-mmio device
that there are new buffers to process in a queue. So, all virtqueues
of this device will listen to this register.

In the case of multiple ioevevtfds with the same address range,
it is necessary to determine which ioeventfd to use through `datamatch` and `data`.

However, according to the description of the rust `slice` library, the action
of function `binary_search_by` is:
    `If there are multiple matches, then any one of the matches could be returned.`

Thus, the virtio-mmio device may encounter various exceptions because of
this random unexpected ioeventfd calling.

Delete this `binary_search_by`.

Fix: 07f2d03883574(AddressSpace: Optimized the ioeventfd mechanism)

Signed-off-by: liuxiangdong <liuxiangdong5@huawei.com>