micro_vm: execute the shutdown action in the main thread
In Hongmeng, the VCPU status is queried when the virtual machine shuts down.
In micro VM, the VCPU is already trapped for executing a shutdown action
and will not change the state of the VCPU, while shutting down will keep
querying for state changes, resulting in a loop timeout. This leads to
the micro vm getting stuck when shutting down on hongmeng. Move the shutdown
action of the micro VM to the main thread, so that the VCPU thread will not
get stuck by itself and the shutdown can proceed normally.

Signed-off-by: liuxiangdong <liuxiangdong5@huawei.com>