Use mlock after mmap make sure backend-file work fine in hmv
In hmv, the memory mmaped by a backend file and be accessed by guest vm may be
swaped, which may cause the level1 page(hva->hpa) table be freed. In this case
the level2 page table(goa->hpa) will not be freed (so far, planned to fix later,
then revert this patch).

Signed-off-by: Xiaohang Hu <huxiaohang1@huawei.com>