Introduce chardev-backed flow-control mechanic for serial port
Before the patch, we had no any flow control on PL011 and UART-serial.
The data was sent synchronously with vcpu-thread.

Switch PL011 and UART to use async mechanic. To to so,
we need correctly report to the guest the state of the TX-FIFO,
whether it is full or empty, and properly generate TX interrupt.

Provide ability to register TX-complete event in the chardev-backend
and profide functions to obtain state of the buffers.

Signed-off-by: Dmitry Skorodumov <skorodumov.dmitry@huawei.com>