Fix bugs when playing video on potplayer
Question I: When a video is played, the progress bar keeps blinking. This is because the progress bar of the PotPlayer is not involved in the overall rendering and is displayed separately. Sometimes, the progress bar is delivered separately from the DWM, and sometimes the progress bar is loaded with the texture of the entire window. The load content is omitted by the blit mode optimization, so the flicker occurs. We should clip the loading in this case.

Question II: The blit optimization is based on whether the soft mouse and hard mouse switchover is enabled. This is useful in most scenarios. However, when a video is played, the mouse disappears after the mouse is placed in the window for a while. In this case, the soft mouse switchover is triggered. In this case, the blit optimization should not be stopped.

Signed-off-by: h957891668 <957891668@qq.com>