io_uring: delete ASYNC for performance
Without ASYNC flag, the operation for io_uring is to try and issue an sqe
as non-blocking first, and if that fails, execute it in an async manner.
With ASYNC flag it ask for an sqe to be issued aysnc form the start.
Note that this flag immediately causes the SQE to be offloaded to an async
helper thread with no initial non-blocking attempt.  This may be less
efficient and should not be used liberally or without understanding the
performance and efficiency tradeoffs.

Signed-off-by: Yan Wang <wangyan122@huawei.com>