qcow2: fix refcount updating error when deleting snapshot
When deleting snapshot, snapshot information in memory will be deleted
firstly.
Therefore, the size of the snapshot table in the updated memory (assuming size1)
is not the size of the old snapshot table (assuming size0). If this `size1` is
used to release the old snapshot table refcount, it may cause problems.

For example, if there is only one snapshot and the updated size is 0, using `0`
to release the old snapshot table will result in no clusters being released,
causing problems with the refcount block count.

Fix this problem.

Signed-off-by: liuxiangdong <liuxiangdong5@huawei.com>