usb_host: Fixed the deadlock problem that may occur with extremely low probability
There is a very low probability that when the xhci-iothread responds to the front-end driver sending a stop_endpoint command, the epoll_wait on the same device also returns and enters the processing flow. In this case, the two threads will hold the locks on xfer and dev respectively and fall into a deadlock.

Signed-off-by: Fei Xu <xufei30@huawei.com>