ohcam: fix reported formats' sequence
This patch is just for some APP. This fucking stupid APP just uses
the format we report first to realize camera-related functions, while
it may not support this format actually.
For this reason, we put YUY2 format on first position. (On OHOS, we
just support YUY2 and NV12 for camera now.)

Signed-off-by: Fei Xu <xufei30@huawei.com>