xhci: Fix possible stuck
According to xHCI Spec 3.2.7/6.4.1/4.9.1, zero-length packet is required
when exact multiple of max packet size is transferred, it is essential
for proper stream termination in bulk/interrupt transfers, skip TRB
submission for zero-byte transfers may end up in Device hanging, waiting
for status phase completion. zero-length packet usually comes with the
default address 0. In this case, the correct way to handle it is to skip
the address translation and leave the iovec empty.

Signed-off-by: Fei Xu <xufei30@huawei.com>