virtio/mtt: introduce device feature
Now we need to be aware of which device type has been initialized by the
guest. Its initialized type might not match what the backend supports.
This patch introdcues device features to indicate what kind of the
device the backend supports and what kind of the device the driver
initializes.

Signed-off-by: Xiaohang Hu <huxiaohang1@huawei.com>