memory: set default maxmem to current memory size
If the virtio mem device is not configured, the actual memory size used
will not increase. If configuring virtio mem devices, the maximum memory
size needs to be specified in the command line.

So, modify the default value of maxmem to be consistent with the current memory,
instead of a fixed value of 256G.

Signed-off-by: liuxiangdong <liuxiangdong5@huawei.com>