aio: fix misaligned dio for thread_pool
Currently, for misaligned dio, it will use thread pool to do the
read/write operation.
There is an error while sending 2 requests for dio write:
  1) Req1: write offset 0, size 512, it will use thead1 to handle it.
  2) Req2: write offset 512, size 512, it will use thread2 to handle it.
Req1 is not aligned to 4k for dio, it will read 4k data at offset 0,
and replace the data in [0, 512) with the data in Req1, and write the
new 4k data to file.
Req2 is also not aligned to 4k for dio, and do the same thing like Req1.
The data in Req1 may be overwritten by Req2.

      thread1                    thread2
        |                           |
    read [0, 4k) data               |
        |                     read [0, 4k) data
   replace [0, 512) data            |
        |                     replace [512, 1024) data
 write [0, 4k) to file              |
        |                     write [0, 4k) to file which overwrite the [0, 512) data
       end                          |
                                   end

Fix bfd97545c7("file: obtain IO alignment size from file system") which aligned dio to 4k
at least.

Signed-off-by: Yan Wang <wangyan122@huawei.com>