chardev_backend: introduce kick_out_evt
It's turned out, that we cannot rely on EventSet::OUT event
after EventLoop::update_event(), since sometimes it races
and event is lost. Before the patch, sometime we can observe
situation when out_handler is finished with buffer empty
and EventSet::OUT is cleared. But in another thread we are
registering listener for EventSet::OUT in fill_outbuf().
But there is no guarantee about order in which register/unregister will
be executed, so EventSet::OUT was not registered and sends
are totally stalled!

To fix the race, use separate event to kick sends and introduce
separate handlers for EventSet::OUT and kick-event.

Signed-off-by: Dmitry Skorodumov <skorodumov.dmitry@huawei.com>